Проведен рефакторинг ExperienceView
