--- src/components/Dialogs/MoreCase.vue
@@ -1,20 +1,0 @@
-<template>
-    <div class="main_content">
-
-        
-
-    </div>
-</template>
-
-<script setup>
-
-</script>
-
-<style scoped>
-
-.main_content {
-    
-}
-
-
-</style>
